PC_Game/requirements.txt
flask==2.0.1
flask-socketio==5.1.1
python-socketio==5.4.0
flask-cors==3.0.10
pygame==2.1.2
